Payment Decimal  Working

+++ BackendSaiKitchen/Controllers/InquiryController.cs
@@ -82,7 +82,7 @@ namespace SaiKitchenBackend.Controllers
             foreach (var payment in inquiry.Payments)
             {
                 payment.PaymentStatusId = (int)paymentstatus.PaymentApproved;
-                Helper.AddPayment((long)payment.PaymentAmount);
+                Helper.AddPayment(payment.PaymentAmount);
                 payment.CreatedDate = Helper.GetDateTime();
                 payment.CreatedBy = Constants.userId;
                 payment.UpdatedBy = Constants.userId;
